:recycle: refactor: Moved circular progress to UI components
